Fixed Edit function
