+++ Feed/index.html
@@ -8,6 +8,7 @@
     <script src="/JS/search.js" type="module" defer></script>
     <script src="/JS/posts.js" type="module" defer></script>
     <script src="/JS/newpost.js" type="module" defer></script>
+    <script src="/JS/editpost.js" type="module" defer></script>
     
     <title>ChatApp - Feed</title>
 </head>
@@ -51,7 +52,7 @@
                         </div>
                         <form id="newPostForm">
                             <div class="card-body">
-                                <textarea class="form-control" id="post-form-title" rows="1" placeholder="Title"></textarea>
+                                <textarea class="form-control" id="post-form-title" rows="1" placeholder="Title" required></textarea>
                                 <textarea class="form-control" id="post-form-body" rows="3" placeholder="Write your post..."></textarea>
                                 <textarea class="form-control" id="post-form-media" rows="1" placeholder="Media (add URL)"></textarea>
                             </div>
@@ -78,6 +79,27 @@
               <a class="dropdown-item" id="filter-comments" href="#">Comments</a>
             </div>
           </div>
+
+        
+          <div id="edit-post-modal" class="modal">
+            <div class="modal-content">
+                <span class="close-button">&times;</span>
+                <h2>Edit Post</h2>
+                <form id="edit-post-form">
+                    <label for="edit-post-title">Title:</label>
+                    <input type="text" id="edit-post-title" required>
+        
+                    <label for="edit-post-body">Body:</label>
+                    <textarea id="edit-post-body"></textarea>
+        
+                    <label for="edit-post-media">Media URL:</label>
+                    <input type="text" id="edit-post-media">
+        
+                    <button type="submit" class="btn btn-primary" id="editPostButton">Save</button>
+                </form>
+            </div>
+        </div>
+        
 
         <div class="container mt-5">
             <div class="row justify-content-center">
